'Database Backup added'

+++ start.sh
@@ -1,6 +1,5 @@
 echo "Installing dependencies..."
 pip3 install -r requirements.txt
-echo "Checking  Backup files"
 if [ -d "database" ]
 then
     echo "Database folder found"
@@ -8,20 +7,25 @@ else
     mkdir "database"
     echo "Database folder created"
 fi
+if [ -d "dbbackup" ]
+then
+    echo "Database backup folder found"
+else
+    mkdir "dbbackup"
+    echo "Database backup folder created"
+fi
 if [ -f "dbbackup/teleBot.sqlite" ]
 then
-    rm "database/teleBot.sqlite"
+    if [ -f "database/teleBot.sqlite" ]
+    then
+        rm "database/teleBot.sqlite"
+    fi
     cp "dbbackup/teleBot.sqlite" "database/teleBot.sqlite"
     echo "Backup Restored"
 elif [ -f "database/teleBot.sqlite" ]
 then
-    mkdir "dbbackup"
-    echo "Backup folder created"
     cp "database/teleBot.sqlite" "dbbackup/teleBot.sqlite"
     echo "Present Database Backup completed"
-else 
-    mkdir "dbbackup"
-    echo "Created Database Backup folder"
 fi
 echo "initialising the bot..."
 python pyFiles/CR_ALT.py
